+++ TidalPy/tides/multilayer/stress_strain.py
@@ -1,4 +1,4 @@
-""" Calculation of stress, strain, and displacement based on tidal solutions and tidal potential
+""" Calculation of stress and strain based on tidal solutions and tidal potential.
 
 References
 ----------
@@ -11,57 +11,17 @@ B13   : Beuthe (2013, DOI: 10.1016/j.icarus.2012.11.020)
 from typing import Tuple
 
 import numpy as np
+from numba import prange
 
-from .decompose import decompose
-from ...utilities.performance import njit
 from ...utilities.types import FloatArray
+from ...utilities.performance import njit
 
-StressType = Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray, np.ndarray, np.ndarray]
-StrainType = Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray, np.ndarray, np.ndarray]
-
-
-@njit(cacheable=True)
-def calculate_displacements(
-    tidal_potential: np.ndarray,
-    tidal_potential_partial_theta: np.ndarray, tidal_potential_partial_phi: np.ndarray,
-    tidal_solution_y: np.ndarray, colatitude: FloatArray
-    ):
-    """ Calculate tidal displacements using the tidal potential and its partial derivatives as well as the y-solution
-    vector.
-
-    Parameters
-    ----------
-    tidal_potential : FloatArray
-        The tidal potential
-    tidal_potential_partial_theta : FloatArray
-        The partial derivative of the tidal potential with respect to the colatitude
-    tidal_potential_partial_phi : FloatArray
-        The partial derivative of the tidal potential with respect to the longitude
-    tidal_solution_y : np.ndarray
-        Vector of tidal solutions often denoted as a lower-case "y"
-    colatitude : FloatArray
-        Co-latitude where the calculations are conducted
-
-    Returns
-    -------
-    radial_displacement : FloatArray
-        The tidal displacement in the radial direction (r)
-    polar_displacement : FloatArray
-        The tidal displacement in the polar direction (theta)
-    azimuthal_displacement : FloatArray
-        The tidal displacement in the azimuthal direction (phi)
-    """
-
-    # TODO: add in summation over l and m?
-    radial_displacement = tidal_solution_y[0, :] * tidal_potential
-    polar_displacement = tidal_solution_y[2, :] * tidal_potential_partial_theta
-    azimuthal_displacement = tidal_solution_y[2, :] * tidal_potential_partial_phi / np.sin(colatitude)
-
-    return radial_displacement, polar_displacement, azimuthal_displacement
+StressType = np.ndarray
+StrainType = np.ndarray
 
 
-# @njit(cacheable=True)
-def calculate_strain_stress_heating(
+@njit(cacheable=True, parallel=True)
+def calculate_strain_stress(
     tidal_potential: np.ndarray,
     tidal_potential_partial_theta: np.ndarray, tidal_potential_partial_phi: np.ndarray,
     tidal_potential_partial2_theta2: np.ndarray, tidal_potential_partial2_phi2: np.ndarray,
@@ -73,6 +33,11 @@ def calculate_strain_stress_heating(
     ) -> Tuple[StrainType, StressType]:
     """ Calculate tidal strain tensor using the tidal potential and its partial derivatives as well as the y-solution
     vector.
+
+    OPT: This function has actually been fairly optimized for performance since it is doing calculations on arrays that
+        are n x m x k x p and can easily reach into the millions of elements. It also must be performed for each
+        tidal mode which could be dozens per time step. It is pretty efficient right now but could always use a
+        second pair of eyes due to its importance.
 
     Parameters
     ----------
@@ -133,7 +98,6 @@ def calculate_strain_stress_heating(
             Stress tensor component - colatitude longitude (appears twice in the symmetric tensor)
 
     """
-
     # Optimizations
     sin_theta = np.sin(colatitude)
     cos_theta = np.cos(colatitude)
@@ -146,42 +110,61 @@ def calculate_strain_stress_heating(
     y2 = tidal_solution_y[1]
     y3 = tidal_solution_y[2]
     y4 = tidal_solution_y[3]
+    y4_shear = y4 / shear_moduli
     dy1_dr = (1. / (lame + 2. * shear_moduli)) * (y2 - (lame * r_inverse) * (2. * y1 - order_l * (order_l + 1.) * y3))
+    y3_r = y3 * r_inverse
+    y1_r = y1 * r_inverse
+    radius_n = y1.shape[0]
+    stress_strain_shape = (6, radius_n, *tidal_potential.shape)
+
+    # Build optimization terms
+    s2_t1 = (1. / sin_theta**2) * tidal_potential_partial2_phi2 + cot_theta * tidal_potential_partial_theta
+    s4_t0 = tidal_potential_partial_phi / sin_theta
+    s5_t0 = 2. * (tidal_potential_partial2_theta_phi - cot_theta * tidal_potential_partial_phi) / sin_theta
 
     # Strain Terms - TB05 Eq. 10; B13 Eq. 8 & 9
     # Build matrix for strain calculations
     # Strain and stress will have a shape of [6x strain/stress terms, radius_N, long_N, lat_N, time_N]
-    strains = np.zeros((6, *y1.shape, *tidal_potential.shape), dtype=np.complex128)
-    # \epsilon_{rr}
-    strains[0, :, :, :, :] = dy1_dr * tidal_potential
-    # \epsilon_{\theta\theta}
-    strains[1, :, :, :, :] = r_inverse * (y3 * tidal_potential_partial2_theta2 + y1 * tidal_potential)
-    # \epsilon_{\phi\phi}
-    # There is a typo in Tobie+2005 with in both the \theta,\phi and \phi\phi components of the strain tensor,
-    #    as pointed out in Kervazo et al (2021; A&A) Appendix D
-    strains[2, :, :, :, :] = r_inverse * (y1 * tidal_potential +
-                                          y3 * ((1. / sin_theta**2) * tidal_potential_partial2_phi2 +
-                                                cot_theta * tidal_potential_partial_theta))
-    # \epsilon_{r\theta}
-    strains[3, :, :, :, :] = y4 * tidal_potential_partial_theta / shear_moduli
-    # \epsilon_{r\phi}
-    strains[4, :, :, :, :] = y4 * tidal_potential_partial_phi / (shear_moduli * sin_theta)
-    # \epsilon_{\theta\phi}
-    strains[5, :, :, :, :] = 2. * r_inverse * (y3 / sin_theta) * \
-                            (tidal_potential_partial2_theta_phi - cot_theta * tidal_potential_partial_phi)
+    strains = np.empty(stress_strain_shape, dtype=np.complex128)
+    stresses = np.empty(stress_strain_shape, dtype=np.complex128)
 
-    # Calculate stress assuming isotropic media (Takeuchi & Saito (1972))
-    # First calculate the trace of strain \epsilon_{rr} + \epsilon_{\theta\theta} + \epsilon_{\phi\phi}
-    strain_trace = strains[0, :, :, :, :] + strains[1, :, :, :, :] + strains[2, :, :, :, :]
+    for ri in prange(radius_n):
+        # Pull out radius dependent parameters
+        dy1dr_, y1_r_, y3_r_, y4_shear_, shear_, lame_ = \
+            dy1_dr[ri], y1_r[ri], y3_r[ri], y4_shear[ri], shear_moduli[ri], lame[ri]
 
-    # Constitutive equation: \sigma_{ij} = 2\bar{\mu} \epsilon_{ij} + \bar{\lambda}\epsilon_{kk}\delta_{ij}
-    stresses = 2. * shear_moduli * strains
-    # \epsilon_{rr}
-    stresses[0, :, :, :, :] += lame * strain_trace
-    # \epsilon_{\theta\theta}
-    stresses[1, :, :, :, :] += lame * strain_trace
-    # \epsilon_{\phi\phi}
-    stresses[2, :, :, :, :] += lame * strain_trace
+        # Any optimizations
+        y1_r_tidal_potential = y1_r_ * tidal_potential
+
+        # \epsilon_{rr}
+        strains[0, ri, :, :, :] = dy1dr_ * tidal_potential
+        # \epsilon_{\theta\theta}
+        strains[1, ri, :, :, :] = y3_r_ * tidal_potential_partial2_theta2 + y1_r_tidal_potential
+        # \epsilon_{\phi\phi}
+        # There is a typo in Tobie+2005 with in both the \theta,\phi and \phi\phi components of the strain tensor,
+        #    as pointed out in Kervazo et al (2021; A&A) Appendix D
+        strains[2, ri, :, :, :] = y1_r_tidal_potential + y3_r_ * s2_t1
+        # \epsilon_{r\theta}
+        strains[3, ri, :, :, :] = y4_shear_ * tidal_potential_partial_theta
+        # \epsilon_{r\phi}
+        strains[4, ri, :, :, :] = y4_shear_ * s4_t0
+        # \epsilon_{\theta\phi}
+        strains[5, ri, :, :, :] = y3_r_ * s5_t0
+
+        # Calculate stress assuming isotropic media (Takeuchi & Saito (1972))
+        # First calculate the trace of strain \epsilon_{rr} + \epsilon_{\theta\theta} + \epsilon_{\phi\phi}
+        strain_trace = strains[0, ri, :, :, :] + strains[1, ri, :, :, :] + strains[2, ri, :, :, :]
+        strain_trace_lame = lame_ * strain_trace
+
+        # Constitutive equation: \sigma_{ij} = 2\bar{\mu} \epsilon_{ij} + \bar{\lambda}\epsilon_{kk}\delta_{ij}
+        stresses[:, ri, :, :, :] = (2. * shear_) * strains[:, ri, :, :, :]
+
+        # \epsilon_{rr}
+        stresses[0, ri, :, :, :] += strain_trace_lame
+        # \epsilon_{\theta\theta}
+        stresses[1, ri, :, :, :] += strain_trace_lame
+        # \epsilon_{\phi\phi}
+        stresses[2, ri, :, :, :] += strain_trace_lame
 
     # # Calculate Tidal Heating
     # volumetric_heating = (frequency / 2.) * (
@@ -198,5 +181,6 @@ def calculate_strain_stress_heating(
     #     tidal_solution_y, tidal_solution_y_derivative, radius, np.zeros_like(radius),
     #     shear_moduli, bulk_moduli, order_l=order_l
     #     )
+
 
     return strains, stresses